errors resolved

+++ src/controllers/books.controller.js
@@ -16,30 +16,30 @@ export async function indexBook(req, res) {
 export async function storeBook(req, res) {
     const { title, author, description } = req.body
     let image
-    if (req.files?.image) {
-        // Upload image to cloudinary
-        const result = await uploadImage(req.files.image.tempFilePath)
-
-        // Delete image from temporary file
-        await fs.rm(req.files.image.tempFilePath)
-
-        // Set image data
-        image = {
-            url: result.secure_url,
-            public_id: result.public_id
-        }
-    }
-
-    // Create new Book
-    const newBook = new Book({
-        title, author, description, image
-    })
-
     try {
+        if (req.files?.image) {
+            // Upload image to cloudinary
+            const result = await uploadImage(req.files.image.tempFilePath)
+
+            // Delete image from temporary file
+            await fs.rm(req.files.image.tempFilePath)
+
+            // Set image data
+            image = {
+                url: result.secure_url,
+                public_id: result.public_id
+            }
+        }
+
+        // Create new Book
+        const newBook = new Book({
+            title, author, description, image
+        })
         // Save book in database
         await newBook.save()
         res.json(newBook)
     } catch (err) {
+        if (req.files?.image) await fs.rm(req.files.image.tempFilePath)
         return res.status(500).json(err.message)
     }
 }
@@ -80,7 +80,7 @@ export async function removeBook(req, res) {
 
         // If it doesn't find it, it returns a 404 error.
         if (!book) return res.sendStatus(404)
-        
+
         // If you have an image, delete it from cloudinary
         if (book.image.public_id) await deleteImage(book.image.public_id)
         res.sendStatus(204)
